+++ Frontend/monarch/lib/utils/overall_colors.dart
