+++ Frontend/monarch/pubspec.yaml
@@ -16,6 +16,7 @@ dependencies:
   fl_chart: ^1.0.0
   google_fonts: ^6.2.1
   shared_preferences: ^2.5.3
+  intl: ^0.20.2
 
 dev_dependencies:
   flutter_test:
